+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## 1.0 branch
 ### 1.0.1-beta2
 * `PhpDocMakerCommand` now correctly handles PHP errors too (e.g. notice);
+* each entity now has a coherent `__toString()` method, as well as a
+    `toSignature()` method that returns a humanized version of its signature;
 * `ClassesExplorer` throws a correct exception on missing Composer autoloader;
 * added `AbstractEntity::getTagsByName()` method. Invalid tags will throw an exception;
 * added `ClassEntity::getParentClass()` method;